isolated entity to just a value object, putting in renderer ftm
