+++ lib/renderer/dom/index.js
@@ -7,7 +7,6 @@ var Interactions = require('./interactions');
 var shallowCopy = require('shallow-copy');
 var virtualize = require('virtualize');
 var virtualize = require('../../virtualize');
-var Entity = require('./entity');
 var each = require('component-each');
 var assign = require('extend');
 var Pool = require('dom-pool');
@@ -15,6 +14,7 @@ var walk = require('dom-walk');
 var zip = require('array-zip');
 var raf = require('component-raf');
 var isDom = require('is-dom');
+var uid = require('get-uid');
 var tree = virtualize.tree;
 
 /**
@@ -146,7 +146,7 @@ function dom(world) {
     // improve the overall performance of their app and avoids hard
     // to track down bugs. We essentially are trading a bit of
     // performance here for user-experience.
-    if (!entity.shouldUpdate()) return next();
+    if (!shouldUpdate(entity)) return next();
 
     // TODO: We really want to just use a different object here
     // whenever the props/state changes so we can just do an equality
@@ -165,7 +165,7 @@ function dom(world) {
     // Now we can commit the state of the entity. All of the
     // pending props and state will now be committed and reflect
     // the actual state of the component.
-    entity.commit();
+    commit(entity);
 
     // Re-render the tree to get an up-to-date representation
     // of the component with the new props/state
@@ -238,7 +238,7 @@ function dom(world) {
    */
 
   function mount(entity) {
-    entity.commit();
+    commit(entity);
     trigger('beforeMount', entity, [entity.props, entity.state]);
 
     // This will store all the entities that are children
@@ -279,7 +279,6 @@ function dom(world) {
     unmountChildren(entity);
     trigger('beforeUnmount', entity, [el, entity.props, entity.state]);
     removeEvents(entity);
-    entity.release();
     delete world.elements[entity.id];
     delete world.renders[entity.id];
     delete world.entities[entity.id];
@@ -837,5 +836,53 @@ function dom(world) {
     if (message) throw new Error(message);
   }
 
+  function commit(entity) {
+    entity.state = entity._pendingState;
+    entity.props = entity._pendingProps;
+    entity._pendingState = assign({}, entity.state);
+    entity._pendingProps = assign({}, entity.props);
+    entity.dirty = false;
+  }
+
+  function shouldUpdate(entity) {
+    if (!entity.dirty) return false;
+    var nextState = entity._pendingState;
+    var nextProps = entity._pendingProps;
+    return entity.component.shouldUpdate(entity.props, entity.state, nextProps, nextState);
+  };
+
   return teardown;
+};
+
+/**
+ * A rendered component instance.
+ *
+ * This manages the lifecycle, props and state of the component.
+ *
+ * @param {Function} component
+ * @param {Object} props
+ */
+
+function Entity(component, props) {
+  this.id = uid();
+  this.options = component.options;
+  this.props = props || {};
+  this.component = component;
+  this.state = this.component.initialState(this.props);
+  this.lifecycle = null;
+  this._pendingProps = assign({}, this.props);
+  this._pendingState = assign({}, this.state);
+  this.dirty = false;
+}
+
+/**
+ * Get an option
+ *
+ * @param {String} name
+ *
+ * @return {*}
+ */
+
+Entity.prototype.option = function(name) {
+  return this.options[name];
 };
